Move fixtures to project wide conftest.py
